==> capstone/semana_4/03_Quiz.ipynb <==
%%html
<h1>Quiz</h1>

<h2>Carga de datos y PCA</h2>

<h3>Pregunta 1</h3>
¿Al hacer la PCA, qué porcentaje de la información retendremos, si nos quedamos con 32 componentes?
<form>
    <input type='radio' name='p_1'>88.1</input></br>
    <input type='radio' name='p_1'>95.3</input></br>
    <input type='radio' name='p_1'>99.9</input></br>
    <input type='radio' name='p_1' checked="true">92.6</input>
</form>

<h3>Pregunta 2</h3>
¿Cuál es el valor màximo de la 1ª variable de los datos etiquetados tras transformarlos mediante PCA?
<form>
    <input type='radio' name='p_2'>1.1</input></br>
    <input type='radio' name='p_2' checked="true">27.5</input></br>
    <input type='radio' name='p_2'>5.3</input></br>
    <input type='radio' name='p_2'>-15.2</input>
</form>

<h2>Regresión Logística</h2>

<h3>Pregunta 3</h3>
¿Cuántas galaxias espirales hay en los datos no etiquetados según el modelo de regresión logística?
<form>
    <input type='radio' name='p_3'>3169</input></br>
    <input type='radio' name='p_3'>6299</input></br>
    <input type='radio' name='p_3'>10000</input></br>
    <input type='radio' name='p_3'>3701</input>
</form>

<h3>Pregunta 4</h3>
Un clasificador que muestre una area de 0.55 por debajo de la curva ROC diremos que es un clasificador 
<form>
    <input type='radio' name='p_4'>bueno</input></br>
    <input type='radio' name='p_4'>espectacular</input></br>
    <input type='radio' name='p_4'>aleatorio</input></br>
    <input type='radio' name='p_4'>perfecto</input>
</form>


<h2>Redes Neuronales</h2>

<h3>Pregunta 4</h3>
¿Cuál es el valor màximo de la 1a variable de los datos etiquetados tras transformarlos mediante PCA?
<form>
    <input type='radio' name='p_1'>0</input></br>
    <input type='radio' name='p_1'>1</input></br>
    <input type='radio' name='p_1'>1</input></br>
    <input type='radio' name='p_1'>21</input>
</form>

<h3>Pregunta 5</h3>
¿Cuál es el valor màximo de la 1a variable de los datos etiquetados tras transformarlos mediante PCA?
<form>
    <input type='radio' name='p_1'>0</input></br>
    <input type='radio' name='p_1'>1</input></br>
    <input type='radio' name='p_1'>1</input></br>
    <input type='radio' name='p_1'>21</input>
</form>
